share federation id definition in client lib
