+++ gateway/cli/src/main.rs
@@ -4,10 +4,9 @@ use bitcoin::{Address, Amount, Transaction};
 use clap::{Parser, Subcommand};
 use fedimint_server::modules::wallet::txoproof::TxOutProof;
 use ln_gateway::{
-    config::GatewayConfig, BalancePayload, DepositAddressPayload, DepositPayload, FederationId,
-    WithdrawPayload,
+    config::GatewayConfig, BalancePayload, DepositAddressPayload, DepositPayload, WithdrawPayload,
 };
-use mint_client::utils::from_hex;
+use mint_client::{utils::from_hex, FederationId};
 use serde::Serialize;
 
 #[derive(Parser)]
